feat: 검색창 store 및 provider 생성 (#37)
