--- src/entities/fontSearch/model/search.store.jsx
@@ -1,16 +1,0 @@
-import { createStore, useStore } from 'zustand'
-
-const initialValues = {
-  isModalOpen: false,
-}
-
-export const modalStore = createStore((set) => ({
-  ...initialValues,
-  actions: {
-    openModal: () => set({ isModalOpen: true }),
-    closeModal: () => set({ isModalOpen: false }),
-  },
-}))
-
-export const useIsModalOpen = () => useStore(modalStore, (state) => state.isModalOpen)
-export const useModalActions = () => useStore(modalStore, (state) => state.actions)
