+++ src/entities/search/model/searchProvider.provider.jsx
@@ -1,0 +1,6 @@
+import { modalStore } from './modal.store'
+import { SearchStoreContext } from './search.store'
+
+export const SearchStoreProvider = ({ children }) => {
+  return <SearchStoreContext.Provider value={modalStore}>{children}</SearchStoreContext.Provider>
+}
